requirements.txt
imageio==2.9.0
numpy==1.18.5
opencv-python==4.3.0.36
pandas==1.1.0
Pillow==7.1.2
scikit-image==0.17.2
scipy==1.5.2
Flask==1.1.2
tqdm==4.48.2
imageio-ffmpeg==0.4.2
torch===1.6.0 
torchvision===0.7.0 -f https://download.pytorch.org/whl/torch_stable.html